Move the ssl utils to shared package
SSL functions and types could be used in other parts than mgradm, move
it to shared.
