+++ REUSE.toml
@@ -7,7 +7,7 @@ SPDX-PackageSupplier = "Uyuni Project <>"
 SPDX-PackageDownloadLocation = "https://github.com/uyuni-project/uyuni-tools"
 
 [[annotations]]
-path = ["mgradm/shared/ssl/testdata/**", "go.mod", "go.sum", "uyuni-tools.changes*", "uyuni-tools.spec", ".tito/**"]
+path = ["shared/ssl/testdata/**", "go.mod", "go.sum", "uyuni-tools.changes*", "uyuni-tools.spec", ".tito/**"]
 precedence = "aggregate"
 SPDX-FileCopyrightText = "2023-2024 SUSE LLC"
 SPDX-License-Identifier = "Apache-2.0"
