+++ shared/ssl/testdata/chain1/intermediate-ca.crt
