+++ shared/ssl/testdata/chain1/root-ca.crt
