+++ shared/ssl/testdata/chain2/RHN-ORG-TRUSTED-SSL-CERT
